test: move test cases to `operator.rs`
