+++ core/src/ast/expr.rs
@@ -183,204 +183,10 @@ mod tests {
             }
             .to_sql()
         );
-
-        assert_eq!(
-            "id - num",
-            Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Minus,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-
-        assert_eq!(
-            "id * num",
-            Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Multiply,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-
-        assert_eq!(
-            "id / num",
-            Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Divide,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-
-        assert_eq!(
-            "id % num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Modulo,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-
-        assert_eq!(
-            "id + num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::StringConcat,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id > num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Gt,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id < num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Lt,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id >= num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::GtEq,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id <= num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::LtEq,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id = num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Eq,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id <> num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::NotEq,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id AND num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::And,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id OR num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Or,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id XOR num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Xor,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id LIKE num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::Like,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id ILIKE num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::ILike,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id NOT LIKE num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::NotLike,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "id NOT ILIKE num",
-            &Expr::BinaryOp {
-                left: Box::new(Expr::Identifier("id".to_string())),
-                op: BinaryOperator::NotILike,
-                right: Box::new(Expr::Identifier("num".to_string()))
-            }
-            .to_sql()
-        );
-        assert_eq!(
-            "+id",
-            Expr::UnaryOp {
-                op: UnaryOperator::Plus,
-                expr: Box::new(Expr::Identifier("id".to_owned())),
-            }
-            .to_sql(),
-        );
-
         assert_eq!(
             "-id",
             Expr::UnaryOp {
                 op: UnaryOperator::Minus,
-                expr: Box::new(Expr::Identifier("id".to_owned())),
-            }
-            .to_sql(),
-        );
-
-        assert_eq!(
-            "NOT id",
-            Expr::UnaryOp {
-                op: UnaryOperator::Not,
-                expr: Box::new(Expr::Identifier("id".to_owned())),
-            }
-            .to_sql(),
-        );
-
-        assert_eq!(
-            "id!",
-            Expr::UnaryOp {
-                op: UnaryOperator::Factorial,
                 expr: Box::new(Expr::Identifier("id".to_owned())),
             }
             .to_sql(),
